moved bootstrap strip content to HadoopInTheCloudRunner
