+++ mrjob/cloud.py
@@ -14,6 +14,7 @@
 # limitations under the License.
 import logging
 import os
+import pipes
 from mrjob.runner import MRJobRunner
 from mrjob.runner import RunnerOptionStore
 from mrjob.util import cmd_line
@@ -34,7 +35,15 @@ class HadoopInTheCloudJobRunner(MRJobRunner):
     # init: mentions
     # _cluster_id
     # _bootstrap
+    # _bootstrap_dir_mgr
     # _master_bootstrap_script_path
+
+    ### Bootstrapping ###
+
+    def _cp_to_local_cmd(self):
+        """Command to copy files from the cloud to the local directory
+        (usually via Hadoop). Redefine this as needed"""
+        return 'hadoop fs -copyToLocal'
 
     def _create_master_bootstrap_script_if_needed(self):
         """Helper for :py:meth:`_add_bootstrap_files_for_upload`.
@@ -99,3 +108,79 @@ class HadoopInTheCloudJobRunner(MRJobRunner):
                 f.write(line)
 
         self._master_bootstrap_script_path = path
+
+    def _master_bootstrap_script_content(self, bootstrap):
+        """Return a list containing the lines of the master bootstrap script.
+        """
+        out = []
+
+        def writeln(line=''):
+            out.append(line + '\n')
+
+        # shebang, precommands
+        self._write_start_of_sh_script(writeln)
+
+        # store $PWD
+        writeln('# store $PWD')
+        writeln('__mrjob_PWD=$PWD')
+        writeln()
+
+        # special case for PWD being in /, which happens on Dataproc
+        # (really we should cd to tmp or something)
+        writeln('if [ $__mrjob_PWD = "/" ]; then')
+        writeln('  __mrjob_PWD=""')
+        writeln('fi')
+        writeln()
+
+        # run commands in a block so we can redirect stdout to stderr
+        # (e.g. to catch errors from compileall). See #370
+        writeln('{')
+
+        # download files
+        writeln('  # download files and mark them executable')
+
+        cp_to_local = self._cp_to_local_cmd()
+
+        # TODO: why bother with $__mrjob_PWD here, since we're already in it?
+        for name, path in sorted(
+                self._bootstrap_dir_mgr.name_to_path('file').items()):
+            uri = self._upload_mgr.uri(path)
+            writeln('  %s %s $__mrjob_PWD/%s' %
+                    (cp_to_local, pipes.quote(uri), pipes.quote(name)))
+            # make everything executable, like Hadoop Distributed Cache
+            writeln('  chmod a+x $__mrjob_PWD/%s' % pipes.quote(name))
+        writeln()
+
+        # run bootstrap commands
+        writeln('  # bootstrap commands')
+        for cmd in bootstrap:
+            # reconstruct the command line, substituting $__mrjob_PWD/<name>
+            # for path dicts
+            line = '  '
+            for token in cmd:
+                if isinstance(token, dict):
+                    # it's a path dictionary
+                    line += '$__mrjob_PWD/'
+                    line += pipes.quote(self._bootstrap_dir_mgr.name(**token))
+                else:
+                    # it's raw script
+                    line += token
+            writeln(line)
+
+        writeln('} 1>&2')  # stdout -> stderr for ease of error log parsing
+
+        return out
+
+    def _write_start_of_sh_script(self, writeln):
+        """Write shebang and pre-commands."""
+        # shebang
+        sh_bin = self._sh_bin()
+        if not sh_bin[0].startswith('/'):
+            sh_bin = ['/usr/bin/env'] + sh_bin
+        writeln('#!' + cmd_line(sh_bin))
+
+        # hook for 'set -e', etc. (see #1549)
+        for cmd in self._sh_pre_commands():
+            writeln(cmd)
+
+        writeln()
