+++ mrjob/dataproc.py
@@ -845,75 +845,6 @@ class DataprocJobRunner(HadoopInTheCloudJobRunner):
         """
         return [parse_setup_cmd(cmd) for cmd in self._opts['bootstrap']]
 
-    def _master_bootstrap_script_content(self, bootstrap):
-        """Create the contents of the master bootstrap script.
-        """
-        out = []
-
-        def writeln(line=''):
-            out.append(line + '\n')
-
-        # shebang
-        sh_bin = self._sh_bin()
-        if not sh_bin[0].startswith('/'):
-            sh_bin = ['/usr/bin/env'] + sh_bin
-        writeln('#!' + cmd_line(sh_bin))
-
-        # unused hook for 'set -e', etc. (see #1549)
-        for cmd in self._sh_pre_commands():
-            writeln(cmd)
-        writeln()
-
-        # store $PWD
-        writeln('# store $PWD')
-        writeln('__mrjob_PWD=$PWD')
-
-        # FYI - mtai @ davidmarin - begin section, mtai had to add this
-        # otherwise initialization didn't work
-        # // kept blowing up in all subsequent invocations of $__mrjob_PWD/
-        writeln('if [ $__mrjob_PWD = "/" ]; then')
-        writeln('  __mrjob_PWD=""')
-        writeln('fi')
-        # FYI - mtai @ davidmarin - end section
-
-        writeln()
-
-        # download files
-        writeln('# download files and mark them executable')
-
-        cp_to_local = 'hadoop fs -copyToLocal'
-
-        for name, path in sorted(
-                self._bootstrap_dir_mgr.name_to_path('file').items()):
-            uri = self._upload_mgr.uri(path)
-
-            output_string = '%s %s $__mrjob_PWD/%s' % (
-                cp_to_local, pipes.quote(uri), pipes.quote(name))
-
-            writeln(output_string)
-            # make everything executable, like Hadoop Distributed Cache
-            writeln('chmod a+x $__mrjob_PWD/%s' % pipes.quote(name))
-        writeln()
-
-        # run bootstrap commands
-        writeln('# bootstrap commands')
-        for cmd in bootstrap:
-            # reconstruct the command line, substituting $__mrjob_PWD/<name>
-            # for path dicts
-            line = ''
-            for token in cmd:
-                if isinstance(token, dict):
-                    # it's a path dictionary
-                    line += '$__mrjob_PWD/'
-                    line += pipes.quote(self._bootstrap_dir_mgr.name(**token))
-                else:
-                    # it's raw script
-                    line += token
-            writeln(line)
-        writeln()
-
-        return out
-
     def get_cluster_id(self):
         return self._cluster_id
